+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 5.3.0 - 2019-11-19
+
+- Adding `getLogInputs` to format or order inputs given to console methods
+
+## 5.2.0 - 2019-11-19
+
+- `logger.debug` logs to `console.debug` if available
+
+## 5.1.0 - 2019-11-19
+
+- Adding a `getDateString` option to return a custom date if ISO 8601 is not wanted
+
+## 5.0.1 - 2019-04-19
+
+- Dependencies update
+
 ## 5.0.0 - 2018-07-04
 
 ### Breaking changes